updating pathing
