+++ crates/opsml_interfaces/src/model/base/interface.rs
@@ -835,11 +835,7 @@ impl ModelInterface {
             ));
         }
 
-        let load_path = path
-            .join(SaveName::OnnxModel.to_string())
-            .with_extension(Suffix::Onnx);
-
-        let sess = OnnxSession::load_onnx_session(py, load_path, kwargs)?;
+        let sess = OnnxSession::load_onnx_session(py, path, kwargs)?;
 
         self.onnx_session
             .as_ref()
