+++ crates/opsml_interfaces/src/model/tensorflow/interface.rs
@@ -639,11 +639,7 @@ impl TensorFlowModel {
             ));
         }
 
-        let load_path = path
-            .join(SaveName::OnnxModel.to_string())
-            .with_extension(Suffix::Onnx);
-
-        let sess = OnnxSession::load_onnx_session(py, load_path, kwargs)?;
+        let sess = OnnxSession::load_onnx_session(py, &path, kwargs)?;
 
         self.onnx_session
             .as_ref()
