+++ crates/opsml_interfaces/src/model/torch/interface.rs
@@ -688,11 +688,7 @@ impl TorchModel {
             ));
         }
 
-        let load_path = path
-            .join(SaveName::OnnxModel.to_string())
-            .with_extension(Suffix::Onnx);
-
-        let sess = OnnxSession::load_onnx_session(py, load_path, kwargs)?;
+        let sess = OnnxSession::load_onnx_session(py, &path, kwargs)?;
 
         self.onnx_session
             .as_ref()
